+++ src/models/pagination.rs
@@ -3,9 +3,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize)]
 pub struct PaginationQuery {
     #[serde(default = "default_page")]
-    pub page: i64,
+    pub page: i32,
     #[serde(default = "default_per_page")]
-    pub per_page: i64,
+    pub per_page: i32,
 }
 
 impl PaginationQuery {
@@ -18,7 +18,7 @@ impl PaginationQuery {
         }
     }
 
-    pub fn offset(&self) -> i64 {
+    pub fn offset(&self) -> i32 {
         (self.page - 1) * self.per_page
     }
 }
@@ -31,18 +31,18 @@ pub struct PaginatedResponse<T> {
 
 #[derive(Debug, Serialize)]
 pub struct PaginationMeta {
-    pub current_page: i64,
-    pub per_page: i64,
+    pub current_page: i32,
+    pub per_page: i32,
     pub total_items: i64,
-    pub total_pages: i64,
+    pub total_pages: i32,
 }
 
 impl PaginationMeta {
-    pub fn new(current_page: i64, per_page: i64, total_items: i64) -> Self {
+    pub fn new(current_page: i32, per_page: i32, total_items: i64) -> Self {
         let total_pages = if total_items == 0 {
             1
         } else {
-            (total_items as f64 / per_page as f64).ceil() as i64
+            (total_items as f64 / per_page as f64).ceil() as i32
         };
 
         Self {
@@ -54,10 +54,10 @@ impl PaginationMeta {
     }
 }
 
-fn default_page() -> i64 {
+fn default_page() -> i32 {
     1
 }
 
-fn default_per_page() -> i64 {
+fn default_per_page() -> i32 {
     10
 }
